Benchmark:

```csharp
[MemoryDiagnoser]
public partial class Benchs
{
    public struct Point
    {
        public double X { get; set; }
        public double Y { get; set; }

        public Point(double x, double y)
        {
            X = x;
            Y = y;
        }
    }

    public static Point TransformBaseline(Matrix3x2 matrix, double x, double y)
    {
        if (matrix.IsIdentity)
        {
            return new Point(x, y);
        }

        return new Point(
            (x * matrix.M11) + (y * matrix.M21) + matrix.M31,
            (x * matrix.M12) + (y * matrix.M22) + matrix.M32);
    }

    public static Point TransformUpdated(Matrix3x2 matrix, double x, double y)
    {
        if (matrix.IsIdentity)
        {
            return new Point(x, y);
        }

        var transformed = Vector2.Transform(new Vector2((float)x, (float)y), matrix);
        return new Point(transformed.X, transformed.Y);
    }

    private static Matrix3x2 _doubleScale = Matrix3x2.CreateScale(2.0f);

    [Benchmark]
    public void Baseline()
    {
        TransformBaseline(_doubleScale, 10, 20);
    }

    [Benchmark]
    public void Updated()
    {
        TransformUpdated(_doubleScale, 10, 20);
    }
}
```

|   Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| Baseline | 3.415 ns | 0.0248 ns | 0.0220 ns |     - |     - |     - |         - |
|  Updated | 1.175 ns | 0.0040 ns | 0.0038 ns |     - |     - |     - |         - |